add TODO list with strategy for prototyping
The overall complexity is very high, because we have
1. unbounded input length to decide which grammar rules apply
2. weights to mitigate that storing all possible matches can be too costly
   and backtracking leads to bad search strategies
3. recursion
4. restrictions on number of cores for comparable performance results.

This means something like "LL(n) + arbitrary recursion".

The list also contains the interface and testing sketch.

2 worst case input are identified:
- 1. "recursive" and "complex regex evaluation"
- 2. dynamic input completion

The first requires prototyping of a weight and "any string" fallback as
potential solution.
The second is solved via providing the caller with timeouts for fetching
data and querying.